+++ packages/terra-card/src/Card.jsx
@@ -12,14 +12,24 @@ const propTypes = {
    * Child Nodes
    */
   children: PropTypes.node,
+  /**
+   * An option used to flex the change to a different style of card for different use cases
+   */
+  variant: PropTypes.oneOf(['default', 'raised']),
+};
+
+const defaultProps = {
+  variant: 'default',
 };
 
 const Card = ({
   children,
+  variant,
   ...customProps
 }) => {
   const cardClassNames = cx([
     'card',
+    variant,
     customProps.className,
   ]);
 
@@ -27,5 +37,6 @@ const Card = ({
 };
 
 Card.propTypes = propTypes;
+Card.defaultProps = defaultProps;
 Card.Body = CardBody;
 export default Card;